moved files into corresponding directories
