+++ data_analysis/plot_spectrum.sh
